+++ build.sh
@@ -58,6 +58,7 @@ cat \
   src/cmd_fleet.sh \
   src/cmd_usage.sh \
   src/cmd_digest.sh \
+  src/cmd_proof.sh \
   src/cmd_memory.sh \
   src/cmd_pack.sh \
   src/cmd_secret.sh \